4.21

+++ controllers/blogs.js
@@ -1,8 +1,10 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/api/blogs', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)  
 })
 
@@ -11,16 +13,41 @@ blogsRouter.post('/api/blogs', async (request, response) => {
     response.status(400).end()
   }
   else{
-    const blog = new Blog(request.body)
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+    const body = request.body
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes:body.likes,
+      user: user._id
+    })
     const result = await blog.save()
+    user.blogs = user.blogs.concat(result._id)
+    await user.save()
     response.status(201).json(result)
   }
 
 })
 
 blogsRouter.delete('/api/blogs/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  const userid = decodedToken.id
+  if (!token || !userid) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const blog = await Blog.findById(request.params.id)
+  if ( blog.user.toString() === userid.toString() ){
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+
 })
 
 blogsRouter.put('/api/blogs/:id', async (request, response,) => {
